Refactor type generators into module
