+++ codegen/smithy-ruby-codegen/src/main/java/software/amazon/smithy/ruby/codegen/DirectedRubyCodegen.java
@@ -42,23 +42,23 @@ import software.amazon.smithy.ruby.codegen.generators.AuthGenerator;
 import software.amazon.smithy.ruby.codegen.generators.ClientGenerator;
 import software.amazon.smithy.ruby.codegen.generators.ConfigGenerator;
 import software.amazon.smithy.ruby.codegen.generators.EndpointGenerator;
-import software.amazon.smithy.ruby.codegen.generators.EnumGenerator;
+import software.amazon.smithy.ruby.codegen.generators.types.EnumGenerator;
 import software.amazon.smithy.ruby.codegen.generators.GemspecGenerator;
 import software.amazon.smithy.ruby.codegen.generators.HttpProtocolTestGenerator;
-import software.amazon.smithy.ruby.codegen.generators.IntEnumGenerator;
 import software.amazon.smithy.ruby.codegen.generators.MiddlewareGenerator;
 import software.amazon.smithy.ruby.codegen.generators.ModuleGenerator;
 import software.amazon.smithy.ruby.codegen.generators.PaginatorsGenerator;
 import software.amazon.smithy.ruby.codegen.generators.ParamsGenerator;
 import software.amazon.smithy.ruby.codegen.generators.SpecHelperGenerator;
 import software.amazon.smithy.ruby.codegen.generators.SteepfileGenerator;
-import software.amazon.smithy.ruby.codegen.generators.StructureGenerator;
-import software.amazon.smithy.ruby.codegen.generators.TypesFileBlockGenerator;
-import software.amazon.smithy.ruby.codegen.generators.UnionGenerator;
+import software.amazon.smithy.ruby.codegen.generators.types.StructureGenerator;
+import software.amazon.smithy.ruby.codegen.generators.types.TypesFileBlockGenerator;
+import software.amazon.smithy.ruby.codegen.generators.types.UnionGenerator;
 import software.amazon.smithy.ruby.codegen.generators.ValidatorsGenerator;
 import software.amazon.smithy.ruby.codegen.generators.VersionGenerator;
 import software.amazon.smithy.ruby.codegen.generators.WaitersGenerator;
 import software.amazon.smithy.ruby.codegen.generators.YardOptsGenerator;
+import software.amazon.smithy.ruby.codegen.generators.types.IntEnumGenerator;
 import software.amazon.smithy.ruby.codegen.middleware.MiddlewareBuilder;
 import software.amazon.smithy.ruby.codegen.rulesengine.AuthSchemeBinding;
 import software.amazon.smithy.ruby.codegen.rulesengine.BuiltInBinding;
