+++ codegen/smithy-ruby-codegen/src/main/java/software/amazon/smithy/ruby/codegen/generators/MiddlewareGenerator.java
@@ -50,7 +50,7 @@ public class MiddlewareGenerator extends RubyGeneratorBase {
     }
 
     @Override
-    String getModule() {
+    protected String getModule() {
         return "Middleware";
     }
